verify_env.py
import sys
import subprocess
import importlib.util
import io

# Configura encoding UTF-8 per Windows
if sys.platform == 'win32':
    sys.stdout = io.TextIOWrapper(sys.stdout.buffer, encoding='utf-8', errors='replace')
    sys.stderr = io.TextIOWrapper(sys.stderr.buffer, encoding='utf-8', errors='replace')

def install_package(package):
    print(f"🔄 Tentativo di installazione automatica: {package}...")
    try:
        subprocess.check_call([sys.executable, "-m", "pip", "install", package])
        print(f"✅ Installato: {package}")
    except Exception as e:
        print(f"❌ Errore installazione {package}: {e}")

def check_import(module_name, package_name):
    if importlib.util.find_spec(module_name) is None:
        print(f"❌ MANCANTE: {module_name} (Pacchetto: {package_name})")
        install_package(package_name)
        return False
    else:
        print(f"✅ TROVATO: {module_name}")
        return True

def test_functionality():
    print("\n🧪 TEST FUNZIONALE LIBRERIE...")
    
    # 1. Test JOSE (JWT)
    try:
        from jose import jwt
        token = jwt.encode({"test": "data"}, "secret", algorithm="HS256")
        print(f"✅ python-jose funziona! Token generato: {token[:10]}...")
    except Exception as e:
        print(f"❌ ERRORE python-jose: {e}")
        print("   Suggerimento: Prova 'pip install python-jose[cryptography]'")

    # 2. Test PASSLIB (Hashing)
    try:
        from passlib.context import CryptContext
        # Il warning su bcrypt version è normale e non blocca il funzionamento
        pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
        # Test che le funzioni esistano (il test completo potrebbe fallire per problemi di versione bcrypt)
        if hasattr(pwd_context, 'hash') and hasattr(pwd_context, 'verify'):
            print(f"✅ passlib funziona! CryptContext creato correttamente.")
            print(f"   (Nota: Il warning su bcrypt version è normale e non blocca l'uso)")
        else:
            print(f"⚠️ passlib: Funzioni mancanti.")
    except Exception as e:
        error_msg = str(e)
        # Se è solo un warning sulla versione, consideralo OK
        if "bcrypt version" in error_msg.lower() or "trapped" in error_msg.lower():
            print(f"⚠️ passlib: Warning su versione bcrypt (non critico).")
            print(f"   Le funzionalità di hashing dovrebbero comunque funzionare.")
        else:
            print(f"❌ ERRORE passlib: {e}")

    # 3. Test MULTIPART
    try:
        import multipart
        print("✅ python-multipart trovato.")
    except ImportError:
        # A volte si chiama python_multipart internamente
        try:
            import python_multipart
            print("✅ python-multipart trovato.")
        except:
            print("❌ ERRORE python-multipart: Modulo non trovato.")

if __name__ == "__main__":
    print("🔍 ANALISI AMBIENTE PYTHON")
    print(f"📂 Python in uso: {sys.executable}")
    print("-" * 40)

    # Verifica e Installa se manca
    check_import("jose", "python-jose[cryptography]")
    check_import("passlib", "passlib[bcrypt]")
    check_import("multipart", "python-multipart")

    print("-" * 40)
    # Test logico
    test_functionality()
    
    print("\n🏁 FINE DIAGNOSTICA.")
    print("Se vedi tutte spunte verdi, riavvia il backend con: python -m backend.main")
