pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "yfinance>=1.0.0",
    "fpdf>=1.7.0",
    "xhtml2pdf>=0.2.11",
    "markdown>=3.5.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.6",
    "ddgs>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "duckduckgo-search>=8.1.1",
]
